third_party/WebKit/LayoutTests/external/wpt/eventsource/format-field-id-null-expected.txt

This is a testharness.js-based test.
FAIL EventSource: U+0000 in id fieldassert_equals: expected "" but got "\0\0"
Harness: the test ran to completion.
